pyproject.toml
[tool.poetry]
name = "cbx-experiments"
version = "0.1.0"
description = ""
authors = ["Philip Hierhager <philip.hierhager@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cbx = "^0.1.6"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"
seaborn = "^0.13.2"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
ipywidgets = "^8.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
